requirements.txt
matplotlib
networkx
numpy
Pillow
pytorch-lightning
pytest
scipy
torch
PyYAML
